requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
albumentations>=1.3.0
segmentation-models-pytorch>=0.3.0
rosreestr2coord>=2.0.0
shapely>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
shapely >=2.0.0